requirements.txt
flask==3.1.1
stupidartnet==1.6.0
pillow==11.2.1
pyserial==3.5
pyusb==1.2.1
